refactor : 에러메세지 view에서 분리
