+++ src/test/java/step2/PrizeTest.java
@@ -3,6 +3,7 @@ package step2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step2.Data.Prize;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
